+++ tester/json/no_inputs.json
@@ -1,10 +1,8 @@
 {
-    "fixtures": {
-        "accounts": [
-            0,
-            0
-        ]
-    },
+    "accounts": [
+        0,
+        0
+    ],
     "steps": [
         {
             "transaction": {
