+++ tester/json/one_transaction_two_outputs.json
@@ -1,10 +1,8 @@
 {
-    "fixtures": {
-        "accounts": [
-            100000,
-            0
-        ]
-    },
+    "accounts": [
+        100000,
+        0
+    ],
     "steps": [
         {
             "transaction": {
