+++ tester/json/two_transactions_one_output.json
@@ -1,10 +1,8 @@
 {
-    "fixtures": {
-        "accounts": [
-            100000,
-            0
-        ]
-    },
+    "accounts": [
+        100000,
+        0
+    ],
     "steps": [
         {
             "transaction": {
