+++ tester/src/lib.rs
@@ -1,8 +1,8 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod accounts;
 pub mod checks;
 pub mod context;
 pub mod error;
-pub mod fixtures;
 pub mod steps;
